Tune correction history

+++ include/search.h
@@ -50,8 +50,8 @@ struct searchinfo {
 	int16_t quiet_history[13][64][64];
 	int16_t capture_history[13][7][64];
 
-	int64_t pawn_correction[2][65536];
-	int64_t non_pawn_correction[2][65536];
+	int16_t pawn_correction[2][65536];
+	int16_t non_pawn_correction[2][65536];
 
 	struct transpositiontable *tt;
 	struct history *history;
